Thermometer
This thermometer was created during an exam (3 hours). If you click on the "+1" or "-1" button, you increase or diminue the temperature. If you click on the "+auto" or "-auto" button, the temperature increases or reduces automatically until the maximum or the minimum value is reached. Also, there are thresholds that you can change.